Fix an issue that caused a wrong encoding
When reading svn information via svnlook, it is important to make the process use the systems's default encoding for its output since the tool will always use that.
Although svn commits are stored with UTF-8 encoding, the encoding of the result of svnlook is not encoded in UTF-8 but rather by the OS default.