Fix issue that falsely marked copy-sources as replaced
Previously there was an issue with how the SvnLook implementation handled the copy-source entries in the SVN change result: It just assumed that an item with a copy-source syntax like (from trunk/vendors/baker/bread.txt:r111) was automatically a replaced item but didn't account for the case where this item actually was just the copy-source of a newly added item.

Added some new unit tests to ensure that the handling is correct.