Exclude items with the commit action type "None"
For the ChangedPathCondition, only changed repository paths should be considered. Items with RepositoryItemAction.None were not changed and are only part of the change-set for informational reasons (e.g. merge source). Therefore, these items are now excluded by default when evaluating a ChangedPathCondition. It's still possible to configure a condition that includes these items by providing an "Action" XML attribute that includes the None value.