Add option to provide detailed information about changed paths.
For some tools it's necessary to have information about the action performed or type of an affected path/file. While this information exists in RepoCop, it was previously not possible to pass it to instructions.
This change introduces two new replacement tokens: #affectedfilesdetailed# and #affectedpathsdetailed# which provide a pipe-separated string containing this information (format: {PATH}|{ACTION}|{KIND}).