Add support for "@" tokens. If "@" is used, the content of the token is written into a file and the token is replaced with the name (and path) of this file
This is esp. useful for CommandLineInstruction and the "affectedfiles" token family. If a CommandLineInstruction used one of the "affectedfiles" tokens, all affected files were written into the command line. If a lot of files were affected by the commit, this could lead to a command line that is larger than what Windows supports and an error would occur.
Now we support the following tokens:

#@affectedfiles#
#@affectedfilesdetailed#
#@affectedpaths#
#@affectedpathsdetailed#

The "@" indicates that the content of the token will not be passed directly in the command line. Instead, the content is written to a file (in the system temp folder) and the path and filename of that file are passed in the command line (with a leading "@" to indicate that this is a file reference)
The filename is built using the name of the token and the revision number. Example: "affectedfilesdetailed-12345"

#19