Cleanup CommandLineInstruction; improve error logging
ExecuteFireAndForgetInstruction should use CreateProcessStartInfo, just like ExecuteWaitForExitInstruction. Options of ProcessStartInfo in which these two methods differ are set afterwards. ExecuteWaitForExitInstruction sets "UseShellExecute=false" because only then redirect of Output and Error streams works.
ExecuteFireAndForgetInstruction sets "UseShellExecute=true". In earlier versions it did not set this property at all, but its default is TRUE in full framework but has changed to FALSE in .Net CORE. Therefore to maintain old behavior now that we moved to .NET 8 it explicitly sets "UseShellExecute=true".

ExecuteWaitForExitInstruction now checks if the exit code and differs from the expected exit code and if so, a warning is logged which includes the called filename and arguments. This should help troubleshooting.
#4